1. UserApplicationService 지원하기 신청/취소(추가) - 현재 모집글 신청 수락된 유저일 경우 다시 지원을 하지 못하게 로직 추가 2. Post - 작성자기준 모집마감,모집마감취소 Api 추가

+++ src/main/java/com/project/dogfaw/acceptance/AcceptanceService.java
@@ -42,6 +42,10 @@ public class AcceptanceService {
         //신청자확인
         userApplicationRepository.findByUserAndPost(applier,post)
                 .orElseThrow(()-> new CustomException(ErrorCode.APPLY_NOT_FOUND));
+        //수락완료 저장 전 현재모집인원 체크
+        if(post.getCurrentMember()>=post.getMaxCapacity()){
+            throw new CustomException(ErrorCode.ACCEPTANCE_PEOPLE_SET_CLOSED);
+        }
         //신청상태 삭제
         userApplicationRepository.deleteByUserAndPost(applier,post)
                 .orElseThrow(()-> new CustomException(ErrorCode.APPLY_NOT_FOUND));
@@ -54,12 +58,14 @@ public class AcceptanceService {
     @Transactional
     /*신청거절하기*/
     public void rejection(Long userId, Long postId, User user) {
+        //신청자
         User applier = userRepository.findById(userId)
                 .orElseThrow(()->new CustomException(ErrorCode.SIGNUP_USERID_NOT_FOUND));
+        //게시글
         Post post = postRepository.findById(postId)
                 .orElseThrow(()->new CustomException(ErrorCode.POST_NOT_FOUND));
+        //작성자
         Long writer = post.getUser().getId();
-
 
         //작성자확인
         if(!writer.equals(user.getId())){
